add verb, add, statusicons, add purposes, add game presets, add opportunity to heal ghouls

+++ Content.Shared/Imperial/Vampire/Components/VampireComponent.cs
@@ -144,7 +144,7 @@ namespace Content.Shared.Imperial.Vampire
         public HashSet<EntityUid> Ghouls = new();
 
         [AutoNetworkedField]
-        public int GhoulQuantity = 0;
+        public float GhoulQuantity = 0;
 
         [AutoNetworkedField]
         public bool InvisibleIsActive = false;
